refactor(storage): split sqlite store responsibilities
