+++ docs/design-docs/module-boundary-naming-audit.md
@@ -2,14 +2,14 @@
 
 Status: maintained verified audit
 Last verified: 2026-07-28
-Verification base: `b290b2049151f6150377425635ad46dd46a9ac8b` plus the Gate 9 naming working tree
-Verification branch: `refactor/naming-boundaries`
+Verification base: `e29ecc8ceba0ca4279187929aba7733ee2619d97` plus the Gate 9 SQLite-store working tree
+Verification branch: `refactor/sqlite-download-store`
 
 ## 結論
 
 附件報告指出的七類問題大多成立，但原始證據已被後續重構取代。以目前工作樹重新量測後，Desktop 已成為實際 UI owner、Core 已 headless、下載佇列、HTTP 與 logging 邊界也已收斂；剩餘主要缺口是 media execution context 仍讀取一個 UI projection，以及 aria2、FFmpeg 與 filesystem 的最終 Infrastructure ownership。
 
-目前仍不得宣告整體重構完成或發布 v1.1.0：Gate 9 logging 已透過 PR #94 整合，命名已在本分支收斂，但 large-owner 工作尚未完成，最終 stacked branch 也尚未進入 `main`。版本唯一來源仍是 `1.0.32`。
+目前仍不得宣告整體重構完成或發布 v1.1.0：Gate 9 logging 與 naming 已分別透過 PR #94、#95 整合；large-owner 工作仍在分支逐項收斂，最終 stacked branch 也尚未進入 `main`。版本唯一來源仍是 `1.0.32`。
 
 ## 可重現基線
 
@@ -28,7 +28,7 @@ pwsh ./script/audit-module-boundaries.ps1 `
 | `DownKyi.Core` | 271 | 18,245 |
 | `src/DownKyi.Domain` | 11 | 681 |
 | `src/DownKyi.Application` | 30 | 1,143 |
-| `src/DownKyi.Infrastructure` | 24 | 3,719 |
+| `src/DownKyi.Infrastructure` | 27 | 3,751 |
 | `src/DownKyi.Desktop` | 317 | 44,083 |
 
 `DownKyi` executable 已降為單一 14 行 bootstrap；Desktop 是最大的產品 owner。行數不能單獨證明設計品質，因此後續仍以 project references、runtime type usage 與 architecture tests 判定責任邊界。
@@ -47,7 +47,7 @@ pwsh ./script/audit-module-boundaries.ps1 `
 | resolved | service contracts 依賴 ViewModel | 0 interfaces | completed by Gate 8 |
 | resolved | custom collection contract | 0 custom collection references; standard read-only wrappers | completed by Gate 8 |
 | resolved | naming and folder taxonomy inconsistent | 4 endpoint/role-scoped duplicate groups, 0 generic names, 0 file/type mismatches | completed in Gate 9 naming branch |
-| P2 | oversized owners | 13 production files above 500 physical lines | confirmed, decreasing |
+| P2 | oversized owners | 12 production files above 500 physical lines | confirmed, decreasing |
 | resolved | logging owner too broad | contracts in Application; 268-line provider plus dedicated Infrastructure sink/buffer/retention/export owners | completed by Gate 9 PR #94 |
 | P1 | AI knowledge environment incomplete | required root/docs structure and reproducible audit scripts now exist | resolved |
 
@@ -160,12 +160,11 @@ File/type mismatch baseline 已由 4 項降為 0。Bilibili JSON DTO 與 NFO XML
 
 附件提供的「檔名必須等於第一個型別」正規表示式會誤判 partial、`.axaml.cs`、多型別 DTO 與 interface companion records。此方案不採用。
 
-## Finding 11: 巨檔與 logging owner
+## Finding 11: 巨檔 owners
 
-2026-07-28 的實際 boundary audit 有 13 個 production files 超過 500 physical lines；`DownloadPipeline`、`DownloadTaskProjectionStore`、`ViewMyFavoritesViewModel`、`ViewPublicationViewModel` 與原 715 行 `ApplicationLogProvider` 已從 allowlist 移除。其餘最優先的手寫 owners：
+2026-07-28 的實際 boundary audit 有 12 個 production files 超過 500 physical lines；`DownloadPipeline`、`DownloadTaskProjectionStore`、`ViewMyFavoritesViewModel`、`ViewPublicationViewModel`、`SqliteDownloadTaskStore` 與原 715 行 `ApplicationLogProvider` 已從 allowlist 移除。SQLite Store 從 928 行降為 447 行，交易/初始化協調、Domain row mapping、讀取/quarantine 與 SQL writes 已分成具名 owner，既有 schema/migration/resume tests 保持不變。其餘最優先的手寫 owners：
 
 - `ViewVideoViewModel.cs` 1,020
-- `SqliteDownloadTaskStore.cs` 928
 - `ViewMySpaceViewModel.cs` 669
 - `AddToDownloadService.cs` 663
